requirements.txt
blinker
cffi
charset-normalizer
click
colorama
cryptography
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
google-generativeai
gunicorn
itsdangerous
Jinja2
MarkupSafe
pdfminer.six
pdfplumber
pillow
pycparser
PyMuPDF
pypdfium2
Werkzeug
python-dotenv
psycopg2-binary
SQLAlchemy
